Only allow administrators to edit data

+++ TournamentMaker.Web/Controllers/FieldsController.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Net;
 using System.Web.Mvc;
+using TournamentReport.App_Start;
 using TournamentReport.Models;
 
 namespace TournamentReport.Controllers
@@ -23,7 +24,7 @@ namespace TournamentReport.Controllers
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Field field = db.Fields.Find(id);
+            var field = db.Fields.Find(id);
             if (field == null)
             {
                 return HttpNotFound();
@@ -32,6 +33,7 @@ namespace TournamentReport.Controllers
         }
 
         // GET: Fields/Create
+        [Authorize(Roles = Constants.AdministratorsRoleName)]
         public ActionResult Create()
         {
             return View();
@@ -41,6 +43,7 @@ namespace TournamentReport.Controllers
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
+        [Authorize(Roles = Constants.AdministratorsRoleName)]
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "Id,Name,Description,Address")]Field field)
         {
@@ -55,13 +58,14 @@ namespace TournamentReport.Controllers
         }
 
         // GET: Fields/Edit/5
+        [Authorize(Roles = Constants.AdministratorsRoleName)]
         public ActionResult Edit(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Field field = db.Fields.Find(id);
+            var field = db.Fields.Find(id);
             if (field == null)
             {
                 return HttpNotFound();
@@ -74,7 +78,8 @@ namespace TournamentReport.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,Name,Description,Address")] Field field)
+        [Authorize(Roles = Constants.AdministratorsRoleName)]
+        public ActionResult Edit([Bind(Include = "Id,Name,Description,Address")]Field field)
         {
             if (ModelState.IsValid)
             {
@@ -86,13 +91,14 @@ namespace TournamentReport.Controllers
         }
 
         // GET: Fields/Delete/5
+        [Authorize(Roles = Constants.AdministratorsRoleName)]
         public ActionResult Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Field field = db.Fields.Find(id);
+            var field = db.Fields.Find(id);
             if (field == null)
             {
                 return HttpNotFound();
@@ -101,11 +107,17 @@ namespace TournamentReport.Controllers
         }
 
         // POST: Fields/Delete/5
-        [HttpPost, ActionName("Delete")]
+        [HttpPost]
+        [ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = Constants.AdministratorsRoleName)]
         public ActionResult DeleteConfirmed(int id)
         {
-            Field field = db.Fields.Find(id);
+            var field = db.Fields.Find(id);
+            if (field == null)
+            {
+                return HttpNotFound();
+            }
             db.Fields.Remove(field);
             db.SaveChanges();
             return RedirectToAction("Index");
