+++ TournamentMaker.Web/Controllers/RoundsController.cs
@@ -2,6 +2,7 @@
 using System.Data.Entity;
 using System.Linq;
 using System.Web.Mvc;
+using TournamentReport.App_Start;
 using TournamentReport.Models;
 
 namespace TournamentReport.Controllers
@@ -22,18 +23,22 @@ namespace TournamentReport.Controllers
         //
         // GET: /Rounds/Details/5
 
-        public ViewResult Details(int id)
+        public ActionResult Details(int id)
         {
-            Round round = db.Rounds.Find(id);
+            var round = db.Rounds.Find(id);
+            if (round == null) return HttpNotFound();
+
             return View(round);
         }
 
         //
         // GET: /Rounds/Create
-
+        [Authorize(Roles = Constants.AdministratorsRoleName)]
         public ActionResult Create(string tournamentSlug)
         {
-            var tournament = db.Tournaments.First(t => t.Slug == tournamentSlug);
+            var tournament = db.Tournaments.FirstOrDefault(t => t.Slug == tournamentSlug);
+            if (tournament == null) return HttpNotFound();
+
             ViewBag.TournamentId = new SelectList(db.Tournaments, "Id", "Name", tournament.Id);
             return View();
         }
@@ -42,6 +47,7 @@ namespace TournamentReport.Controllers
         // POST: /Rounds/Create
 
         [HttpPost]
+        [Authorize(Roles = Constants.AdministratorsRoleName)]
         public ActionResult Create(Round round, string tournamentSlug)
         {
             if (ModelState.IsValid)
@@ -57,10 +63,12 @@ namespace TournamentReport.Controllers
 
         //
         // GET: /Rounds/Edit/5
-
+        [Authorize(Roles = Constants.AdministratorsRoleName)]
         public ActionResult Edit(int id)
         {
-            Round round = db.Rounds.Find(id);
+            var round = db.Rounds.Find(id);
+            if (round == null) return HttpNotFound();
+
             ViewBag.TournamentId = new SelectList(db.Tournaments, "Id", "Name", round.TournamentId);
             return View(round);
         }
@@ -69,6 +77,7 @@ namespace TournamentReport.Controllers
         // POST: /Rounds/Edit/5
 
         [HttpPost]
+        [Authorize(Roles = Constants.AdministratorsRoleName)]
         public ActionResult Edit(Round round, string tournamentSlug)
         {
             if (ModelState.IsValid)
@@ -83,20 +92,25 @@ namespace TournamentReport.Controllers
 
         //
         // GET: /Rounds/Delete/5
-
+        [Authorize(Roles = Constants.AdministratorsRoleName)]
         public ActionResult Delete(int id)
         {
-            Round round = db.Rounds.Find(id);
+            var round = db.Rounds.Find(id);
+            if (round == null) return HttpNotFound();
             return View(round);
         }
 
         //
         // POST: /Rounds/Delete/5
 
-        [HttpPost, ActionName("Delete")]
+        [HttpPost]
+        [ActionName("Delete")]
+        [Authorize(Roles = Constants.AdministratorsRoleName)]
         public ActionResult DeleteConfirmed(int id, string tournamentSlug)
         {
-            Round round = db.Rounds.Find(id);
+            var round = db.Rounds.Find(id);
+            if (round == null) return HttpNotFound();
+
             db.Rounds.Remove(round);
             db.SaveChanges();
             return RedirectToAction("Standings", "Home", new {tournamentSlug});
